Move amd_family_id and amd_model_id to src/cpu/amd/src/model.rs
