+++ src/cpu/amd/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "cpu"
+version = "0.1.0"
+authors = ["Oreboot Authors"]
+edition = "2018"
+
+[dependencies]
+arch = { path = "../../arch/x86/x86_64"}
+raw-cpuid = "9.0.0"
+
+[profile.release]
+opt-level = 'z'  # Optimize for size.
