+++ src/cpu/amd/lib.rs
@@ -1,0 +1,4 @@
+
+#![no_std]
+#![deny(warnings)]
+pub mod model;
