+++ src/mainboard/amd/romecrb/Cargo.toml
@@ -8,6 +8,7 @@ edition = "2018"
 arch = { path = "../../../arch/x86/x86_64"}
 clock = { path = "../../../drivers/clock"}
 console = { path = "../../../console" }
+cpu = { path = "../../../cpu/amd" }
 model = { path = "../../../drivers/model" }
 print = { path = "../../../lib/print" }
 payloads = { path = "../../../../payloads"}
@@ -16,12 +17,12 @@ util = { path = "../../../lib/util" }
 wrappers = { path = "../../../drivers/wrappers"}
 
 heapless = "0.4.x"
+raw-cpuid = "9.0.0"
 register = "0.3.2"
 static-ref = "0.1.1"
 postcard = "0.4.3"
-x86_64 = "0.12.2"
 vcell = "0.1.2"
-raw-cpuid = "9.0.0"
+x86_64 = "0.12.2"
 
 [dependencies.uart]
 path = "../../../drivers/uart"
